requirements.txt
scrapling
playwright
duckduckgo-search
requests
httpx
google-generativeai
cloudscraper
pdfplumber
rich
tenacity
python-dotenv
loguru
fake-useragent
tldextract
typer
brotli
